Create type stubs for envparse
Omits support for `force` and `preprocessor` due to the complexities that
they introduce into type annotations (see comments at top of __init__.pyi).

Omits support for a default value of `None` for lists, tuples and dicts,
again due to the complexity this would introduce to the type annotations.
It is arguably better to use an empty list/tuple/dict in these cases
regardless, as then there is only one possible value that means "empty"
rather than two.